Move UserSettings component
