+++ src/pages/lessons/MainLessonView.tsx
@@ -10,7 +10,7 @@ import Material from "../../components/Material";
 import TypedText from "../../components/TypedText";
 import Scores from "../../components/Scores";
 import StrokeTip from "./components/StrokeTip";
-import UserSettings from "../../components/UserSettings";
+import UserSettings from "./components/UserSettings";
 import AussieDictPrompt from "../../components/LessonPrompts/AussieDictPrompt";
 import SedSaidPrompt from "../../components/LessonPrompts/SedSaidPrompt";
 import WordBoundaryErrorPrompt from "../../components/LessonPrompts/WordBoundaryErrorPrompt";
