+++ src/pages/lessons/components/Finished.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import LessonCanvasFooter from "./LessonCanvasFooter";
 import FinishedZeroAndEmptyStateMessage from "./FinishedZeroAndEmptyState";
-import UserSettings from "../../../components/UserSettings";
+import UserSettings from "./UserSettings";
 import {
   stitchTogetherLessonData,
   transformLessonDataToChartData,
